chore: tidy router handler imports
