+++ V2_COMPLIANCE_PROGRESS_TRACKER.md
@@ -162,6 +162,13 @@
 - **Completion Date**: 2025-08-24
 - **Summary**: Brought down from 892 to 148 lines by modularizing handshake, routing, authentication, and logging utilities. Protocol file now orchestrates these modules.
 
+### MOD-017: Frontend Routing System ✅
+- **File**: `src/web/frontend/frontend_router.py`
+- **Status**: Completed
+- **Assigned To**: Current Agent
+- **Completion Date**: 2025-08-24
+- **Summary**: Refactored monolithic router into modular core, middleware, handlers and configuration modules coordinated by an orchestrator. Imports updated and routing tests pass.
+
 ## 📋 AVAILABLE CONTRACTS FOR CLAIMING
 
 > **Note:** Remaining contract descriptions include current line counts for context only. There is no strict LOC target—deliver clean, production-ready, tested modules that honor SRP and SOLID principles.
