+++ src/web/frontend/testing/assertion_helpers.py
@@ -2,7 +2,7 @@
 from typing import Dict, Any
 
 from ..frontend_app import UIComponent
-from ..frontend_router import RouteConfig, NavigationState
+from ..frontend_router_config import RouteConfig, NavigationState
 
 
 class AssertionHelpers:
