+++ src/web/frontend/testing/ui_utils.py
@@ -2,7 +2,7 @@
 from datetime import datetime
 
 from ..frontend_app import UIComponent, create_component
-from ..frontend_router import RouteConfig, NavigationState
+from ..frontend_router_config import RouteConfig, NavigationState
 
 
 class UIInteractionUtilities:
